1-synthetic_cdr_generator/requirements.txt
anyio==4.9.0
attrs==25.3.0
Authlib==1.5.2
avro==1.12.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
confluent-kafka==2.10.0
cryptography==44.0.2
fastavro==1.10.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
numpy==2.2.5
pycparser==2.22
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
sniffio==1.3.1
typing_extensions==4.13.2
Unidecode==1.4.0
urllib3==2.4.0
